+++ scripts/utils/index.js
@@ -2,7 +2,7 @@
 export { CacheManager } from './cacheManagement.js';
 
 // Document Handling
-export { getDocuments, findSelectedDocument } from './documents.js';
+export { fetchDocuments, prepareDocuments } from './documents.js';
 
 // Dropdown Handling
 export { initializeDropdown, updateAbilityDropdowns, generateDropdownHTML } from './dropdowns.js';
@@ -10,9 +10,6 @@ export { initializeDropdown, updateAbilityDropdowns, generateDropdownHTML } from
 // UI Manipulations
 export { registerButton } from './htmlManipulations.js';
 export { handleAbilitySelectionChange, addAbilitySelectionListeners } from './listeners.js';
-
-// Registrations
-export { registerRaces, registerClasses, registerBackgrounds } from './registrations.js';
 
 // Miscellaneous
 export { statRoller } from './statRoller.js';